+++ CHANGELOG.md
@@ -6,3 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+### Added
+
+- Add command `/rules` to show a groups rules
+- Add command `/rules_define YOUR_RULES` to define new rules in a group
+- Add command `/rules_remvoe` to remove the groups rules